+++ RealmSwift/Impl/ComplexTypes.swift
@@ -312,13 +312,18 @@ extension RawRepresentable where Self: _OptionalPersistable, RawValue: _Optional
         if prop.optional {
             prop.swiftAccessor = BridgedPersistedPropertyAccessor<Optional<Self>>.self
         } else {
-            prop.swiftAccessor = PersistedEnumAccessor<Self>.self
+            prop.swiftAccessor = BridgedPersistedPropertyAccessor<Self>.self
         }
     }
 }
 
-extension PersistableEnum {
+// Projections need to be usable in collections, which means they have to be
+// SchemaDiscoverable even though they can't be used with @Persisted
+extension Projection: SchemaDiscoverable {
+    public static var _rlmType: PropertyType {
+        fatalError("Projection types cannot be used with @Persisted and must use @Projected")
+    }
     public static func _rlmDefaultValue(_ forceDefaultInitialization: Bool) -> Self {
-        return self.allCases.first!
+        fatalError()
     }
 }